# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SolidFireGateway, Version=1.1.1.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/SolidFire/{cluster}/snapshots": {
      "post": {
        "tags": [
          "Snapshot"
        ],
        "parameters": [
          {
            "name": "cluster",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSnapshotParams"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSnapshotParams"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSnapshotParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/SolidFire/{cluster}/snapshots/group": {
      "post": {
        "tags": [
          "Snapshot"
        ],
        "parameters": [
          {
            "name": "cluster",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupSnapshotParams"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupSnapshotParams"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupSnapshotParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/SolidFire/{cluster}/snapshots/group/list": {
      "post": {
        "tags": [
          "Snapshot"
        ],
        "parameters": [
          {
            "name": "cluster",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/SolidFire/{cluster}/snapshots/list": {
      "post": {
        "tags": [
          "Snapshot"
        ],
        "parameters": [
          {
            "name": "cluster",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListSnapshotsParams"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ListSnapshotsParams"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ListSnapshotsParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateGroupSnapshotParams": {
        "type": "object",
        "properties": {
          "volumes": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "retention": {
            "type": "string",
            "nullable": true
          },
          "attributes": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateSnapshotParams": {
        "type": "object",
        "properties": {
          "volumeID": {
            "type": "integer",
            "format": "int32"
          },
          "retention": {
            "type": "string",
            "nullable": true
          },
          "attributes": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ListSnapshotsParams": {
        "type": "object",
        "properties": {
          "volumeID": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}